[cudamapper] Refactor word_size to be an argument to scoring function and allow a zero-penalty score for gapless matches.
Word_size (i.e., the kmer size when used as the number of matched basepairs
during scoring) was previously hardcoded in the scoring function. This
has now been refactored as an argument as we prepare for allowing different kmer sizes.

In addition, minimap2 has a zero-cost gap penalty for adjacent matches. This is
now implemented explicitly in overlapper_minimap::log_linear_anchor_weight.