[overlapper_minimap] Fix global indexes using chain_anchors_in_block.
Tweaks the approach in Guo to use a one-block-per-anchor approach and fixes
many off by 1 / off by 64 errors in the global read/write indexes in the function.

The key insight was that each PE in the Guo paper computes the overlaps of 1 read,
and a PE is mapped to one CUDA block, followed by realizing that the "i" and
"counter" variables map to the window starts/ends for a single block's window.

There are still improvements to be made here.
- If we provide read starts/ends, we can perform one tile per read. We would still need to allow multiple chains
per read.
- We can completely eliminate the traceback in
produce_chains by better tracking the first anchor on a chain, the last anchor on a chain, and the number
of anchors within the chain,  making the "traceback" O(1).
- We do not currently produce the correct scores, probably due to global contention for the scores array. Inserting
a compare-and-swap in the write line would probably alleviate this.